+++ topiary/__init__.py
@@ -173,7 +173,7 @@ from .amino_acids import (
     encode_amino_acids,
 )
 
-__version__ = "5.56.1"
+__version__ = "5.56.2"
 
 __all__ = [
     "normalize_python_types",